feat(metadata-sidebar): Reorganize stories
